refactor: use named functions for top-down view
